Add filters back in, SVG-style.
This patch adds in the filter checkboxes (again), but also deals
with them in a nicer way than before. Where once it was only per-
user, now you can link to a filtered chart, because the options
are stored in the URL. There's also an extra filter option. Fun
stuff!

Change-Id: Ib3e8db9e1e9e83d08b108a93557c1070ebd7a183