Fix a few bugs in print mode, and clean whitespace errors
Most of the whitespace errors should be gone now, thanks to Geany,
and the print mode now works properly while zoomed in. Before it
wouldn't find more than two levels of child nodes for whatever
reason (I know the reason, of course: I did it wrong), but now it
will properly show all of the children of the current zoom level.

Change-Id: I6719a030c5d30a112e8db0b70f51e3a7cdad193e