Add in compensation field
This makes use of the (now more functional) user.canSeePrivateData
field, and give so-privileged users the opportunity to see and
edit the compensation of a node. Note: If you get past my meagre
defences and manage to change your pay, you can bet HR won't take
that as a binding contract.

Change-Id: Idbfb6763db3db45798dc0caaa5bf37f617aacc40