Fix node counts
The addToDocCount function wasn't working for me. The new version
actually checks to see how many nodes are in the document, every
time a node is added or removed. Now you should have accurate
numbers whenever you see the front page.

Of course, first you'll need to add or remove a node from the
document.

But that's not _too_ hard.

Change-Id: I0b03cbd3e5246a030df58d3423e1a067b7749feb